+++ DreamNekos.API/Response/Profile/GetProfileResponse.cs
@@ -7,7 +7,7 @@ namespace DreamNekos.API.Response.Profile
     public class GetProfileResponse
     {
         public required Guid Id { get; set; }
-        public required int TgId { get; set; }
+        public required long TgId { get; set; }
         public string? About { get; set; }
         public List<UserInterestLinkResponse> Interest { get; set; } = new List<UserInterestLinkResponse>();
         public List<LinkResponse> Links { get; set; } = new List<LinkResponse>();
